Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
 void main(){
-    int a;
+    float a;
     scanf("%f",&a);
     printf("%f",a);}
